Move SDFG helper for the builder in the subfolder
